+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix: demo image path in README
 - refactor: replace app_version meta tag with `__APP_VERSION__` constant
 - refactor: enforce PascalCase for class names and camelCase for methods and functions
+- refactor: enforce prefer-const
 
 ## 0.12.9 (2024-09-23)